requirements.txt
Flask==3.0.3
gunicorn==23.0.0
Werkzeug==3.0.3
python-dotenv
numpy
pandas
scikit-learn
matplotlib
gensim
openai
gradio
tiktoken
faiss-cpu
datasets
sentencepiece
matplotlib
google-generativeai
scikit-learn
unstructured
plotly
jupyter-dash
pydub
accelerate
sentencepiece
bitsandbytes
psutil
setuptools
speedtest-cli
sentence_transformers
feedparser

